Move adapters inside ports to be more conside with file structure
