+++ server/src/ports/adapters/mongoDbClient.js
@@ -1,6 +1,6 @@
 const MongoClient = require("mongodb").MongoClient;
-const logger = require("../helpers/logger");
-const { sleep } = require("../helpers/utils");
+const logger = require("../../helpers/logger");
+const { sleep } = require("../../helpers/utils");
 
 let client;
 let db;
